Restructure webpack plugin
