+++ webpack-plugin/src/assets/client.html
