+++ webpack-plugin/src/assets/main.css
@@ -1,0 +1,1 @@
+._3qb9NsevMQjxvcRtn0pGaJ{position:absolute;position:fixed;left:0;top:0;bottom:0;width:15em;background:#282830;background:-webkit-linear-gradient(top,#282830,#26262e);background:linear-gradient(180deg,#282830 0,#26262e);filter:progid:DXImageTransform.Microsoft.gradient(startColorstr='#282830',endColorstr='#26262e',GradientType=0);z-index:1;box-shadow:inset -10px 0 10px -4px rgba(0,0,0,.25)}._2xB-CVsT8mGmzTvmKX1J7Q{display:block;padding:2rem 1rem;border-bottom:1px solid rgba(0,0,0,.25);margin:0;text-decoration:none}.iljMtVf66dJmlzOjcgazY{color:#fff;font-family:Helvetica,Arial,sans-serif;font-weight:300;margin:0;padding:0}._27a9H6DrX53JX_u6AJU_Ky{width:100%;height:1.75rem;font-size:.75rem;margin:0;border:none;border-bottom:1px solid rgba(0,0,0,.25);background-color:#42424a;color:#fff;padding:0 1rem}._27a9H6DrX53JX_u6AJU_Ky::-webkit-input-placeholder{color:hsla(0,0%,100%,.5)}._27a9H6DrX53JX_u6AJU_Ky:-moz-placeholder,._27a9H6DrX53JX_u6AJU_Ky::-moz-placeholder{color:hsla(0,0%,100%,.5)}._27a9H6DrX53JX_u6AJU_Ky:-ms-input-placeholder{color:hsla(0,0%,100%,.5)}._27HY8XFRtwEBz20jzrgahq{overflow-y:auto}.iRPsVQOYqgDEVT-8KwPlX{color:hsla(0,0%,100%,.5);background:#282830;background:-webkit-linear-gradient(top,#282830,#26262e);background:linear-gradient(180deg,#282830 0,#26262e);filter:progid:DXImageTransform.Microsoft.gradient(startColorstr='#282830',endColorstr='#26262e',GradientType=0);font-family:-apple-system,BlinkMacSystemFont,Segoe UI,Roboto,Oxygen,Ubuntu,Cantarell,Fira Sans,Droid Sans,Helvetica Neue,sans-serif;text-decoration:none;padding:.5rem 1rem;display:block;font-size:.9rem;font-weight:300;letter-spacing:.05em;text-overflow:ellipsis;overflow:hidden;border-top:1px solid hsla(0,0%,100%,.05);border-bottom:1px solid rgba(0,0,0,.25)}.iRPsVQOYqgDEVT-8KwPlX:hover{color:#fff}.iRPsVQOYqgDEVT-8KwPlX:focus{outline:none;border-left:.15rem solid #4885f6}._2qm4ITmLNIXXHOi4UkmXMH{color:#fff;background:#17181c;border-left:.15rem solid #4885f6;border-top-color:transparent}.iRPsVQOYqgDEVT-8KwPlX:last-child._2qm4ITmLNIXXHOi4UkmXMH{border-bottom:1px solid hsla(0,0%,100%,.05)}._3FmrsCDkkTEEIHeWpQGvbv{background:#17181c;border-left:.15rem solid #4885f6}._2tv6OrrvtME0N5wfXmFkQx{color:hsla(0,0%,100%,.5);border-top-color:transparent;font-family:-apple-system,BlinkMacSystemFont,Segoe UI,Roboto,Oxygen,Ubuntu,Cantarell,Fira Sans,Droid Sans,Helvetica Neue,sans-serif;text-decoration:none;padding:.5rem 1rem .5rem 1.6rem;display:block;font-size:.7rem;font-weight:300;letter-spacing:.05em;text-overflow:ellipsis;overflow:hidden;border-bottom:1px solid rgba(0,0,0,.25);white-space:nowrap}._2tv6OrrvtME0N5wfXmFkQx:hover{color:#fff}._2tv6OrrvtME0N5wfXmFkQx:focus{outline:none;border-left:.15rem solid #a7c6ff}._61xoTsQS3KSpS9LIwruip{color:#fff;background:#17181c;border-left:.15rem solid #a7c6ff;border-top-color:transparent}._5Btbo2K1k_Egnyzl_EVH{position:absolute;bottom:-13.5em;left:0;right:0;height:15em;transform:translateY(0);transition:transform .15s ease-in-out;border-top:1px solid hsla(0,0%,100%,.05)}._3tLC1putZAUja3x4VZY4s2{transform:translateY(-13.5em)}.qzhSsA82etzbyi0DHFh_T{width:50%;border:none;background-color:transparent;height:1.5em;font-size:1em;color:hsla(0,0%,100%,.5);cursor:pointer;padding:0;border-right:1px solid rgba(0,0,0,.25);border-bottom:1px solid rgba(0,0,0,.25);border-top:1px solid transparent;background:linear-gradient(180deg,#282830 0,#26262e);outline:none}._14HBTKtvGdQEs3gVtOaNIb{color:#fff;background:#17181c;border-top:1px solid rgba(0,0,0,.25)}.qzhSsA82etzbyi0DHFh_T:hover{color:hsla(0,0%,100%,.75)}.vTGZLV90FHC6t7FvyRO_T{color:#fff;height:13.5em;position:absolute;top:1.5em;font-family:-apple-system,BlinkMacSystemFont,Segoe UI,Roboto,Oxygen,Ubuntu,Cantarell,Fira Sans,Droid Sans,Helvetica Neue,sans-serif;font-size:1em;padding:.5em;width:100%}._1caLfbBGlUKEc7U1GdmnfU{font-size:1em;font-weight:700;text-align:center;margin:.5em 0}.tb9oCEeBIWQ44HTzZDIIf{width:100%;height:100%}._2PaInoDoPHoRN8rVv8EJtc{background-color:hsla(0,0%,100%,.25);padding:.25em .5em;border-radius:3px;font-size:1.25em;margin-right:.5em}._3ifebI9OJ5UiaFk5aQZAji{width:100%;margin:1em 0;font-size:.8em;line-height:1.5em}._3ifebI9OJ5UiaFk5aQZAji:first-child{margin-top:.5em}._2Ufe2HC8F_iC7F3dB3CU86{color:#fff;font-weight:700;text-decoration:none}._17AQbqzmrXNxYMT53BvxK-{position:absolute;position:fixed;top:0;left:0;cursor:pointer;z-index:2;padding:.5rem;margin:0;background-color:transparent;border:none;font-size:1rem;transition:transform .15s ease-in-out}.Az731RQ_Thf5a-TzhNekQ{fill:#282830}._1JZzEv8q3AnYOlTySBmFeA{transform:translateX(15em)}body,html{height:100%;width:100%;margin:0;padding:0}*,:after,:before{box-sizing:border-box}#styleguide-root,._2XazEebEs3USYsXgJC120e{width:100%;height:100%}.xsfsUu2-1_dTXEr9dLF6G{width:100%;float:right;font-family:Helvetica,Arial,sans-serif;height:100%;z-index:2;position:absolute;right:0;top:0;bottom:0;left:0;background-color:#f4f7f9;transition:transform .15s ease-in-out,width .15s ease-in-out;position:relative}.jsKl8f0kutwVbFxJsyJCr{width:calc(100% - 15em)}._3U-nVzWWpaJ9h3MCIGiGGR{font-weight:300;text-align:center}
