+++ webpack-plugin/src/assets/placeholder.js
