+++ webpack-plugin/src/dynamic-resolve.js
