+++ webpack-plugin/src/loaders/examples-loader.js
