+++ webpack-plugin/src/loaders/plugins-loader.js
